// chrome/browser/resources/settings/chromeos/multidevice_page/multidevice_constants.ts
export enum MultiDeviceSettingsMode {
  NO_ELIGIBLE_HOSTS = 0,
  NO_HOST_SET = 1,
  HOST_SET_WAITING_FOR_SERVER = 2,
  HOST_SET_WAITING_FOR_VERIFICATION = 3,
  HOST_SET_VERIFIED = 4,
}

/**
 * Enum of MultiDevice features. Note that this is copied from (and must
 * include an analog of all values in) the Feature enum in
 * //chromeos/ash/services/multidevice_setup/public/mojom/multidevice_setup.mojom.
 */
export enum MultiDeviceFeature {
  BETTER_TOGETHER_SUITE = 0,
  INSTANT_TETHERING = 1,
  MESSAGES = 2,
  SMART_LOCK = 3,
  PHONE_HUB = 4,
  PHONE_HUB_NOTIFICATIONS = 5,
  PHONE_HUB_TASK_CONTINUATION = 6,
  WIFI_SYNC = 7,
  ECHE = 8,
  PHONE_HUB_CAMERA_ROLL = 9,
}

/**
 * Possible states of MultiDevice features. Note that this is copied from (and
 * must include an analog of all values in) the FeatureState enum in
 * //chromeos/ash/services/multidevice_setup/public/mojom/multidevice_setup.mojom.
 */
export enum MultiDeviceFeatureState {
  PROHIBITED_BY_POLICY = 0,
  DISABLED_BY_USER = 1,
  ENABLED_BY_USER = 2,
  NOT_SUPPORTED_BY_CHROMEBOOK = 3,
  NOT_SUPPORTED_BY_PHONE = 4,
  UNAVAILABLE_NO_VERIFIED_HOST = 5,
  UNAVAILABLE_INSUFFICIENT_SECURITY = 6,
  UNAVAILABLE_SUITE_DISABLED = 7,
  FURTHER_SETUP_REQUIRED = 8,
  UNAVAILABLE_TOP_LEVEL_FEATURE_DISABLED = 9,
  UNAVAILABLE_NO_VERIFIED_HOST_CLIENT_NOT_READY = 10,
  UNAVAILABLE_NO_VERIFIED_HOST_NO_ELIGIBLE_HOST = 11,
}

/**
 * Possible states of Phone Hub's feature access. Access can be
 * prohibited if the user is using a work profile on their phone on Android
 * version <N, or if the policy managing the phone disables access.
 */
export enum PhoneHubFeatureAccessStatus {
  PROHIBITED = 0,
  AVAILABLE_BUT_NOT_GRANTED = 1,
  ACCESS_GRANTED = 2,
}

/**
 * Possible reasons for Phone Hub's feature access being prohibited.
 * Users should ensure feature access is actually prohibited before
 * comparing against these reasons.
 */
export enum PhoneHubFeatureAccessProhibitedReason {
  UNKNOWN = 0,
  WORK_PROFILE = 1,
  DISABLED_BY_PHONE_POLICY = 2,
}
// ...
export interface MultiDevicePageContentData {
  mode: MultiDeviceSettingsMode;
  hostDeviceName: string|undefined;
  betterTogetherState: MultiDeviceFeatureState;
  instantTetheringState: MultiDeviceFeatureState;
  messagesState: MultiDeviceFeatureState;
  smartLockState: MultiDeviceFeatureState;
  phoneHubState: MultiDeviceFeatureState;
  phoneHubCameraRollState: MultiDeviceFeatureState;
  phoneHubNotificationsState: MultiDeviceFeatureState;
  phoneHubTaskContinuationState: MultiDeviceFeatureState;
  phoneHubAppsState: MultiDeviceFeatureState;
  wifiSyncState: MultiDeviceFeatureState;
  isAndroidSmsPairingComplete: boolean;
  cameraRollAccessStatus: PhoneHubFeatureAccessStatus;
  notificationAccessStatus: PhoneHubFeatureAccessStatus;
  appsAccessStatus: PhoneHubFeatureAccessStatus;
  notificationAccessProhibitedReason: PhoneHubFeatureAccessProhibitedReason;
  isNearbyShareDisallowedByPolicy: boolean;
  isPhoneHubPermissionsDialogSupported: boolean;
  isCameraRollFilePermissionGranted: boolean;
  isPhoneHubFeatureCombinedSetupSupported: boolean;
  isChromeOSSyncedSessionSharingEnabled: boolean;
  isLacrosTabSyncEnabled: boolean;
}